refactor(bench): BenchmarkTurn を frame_bench_test.go へ統合する
新規ファイル turn_bench_test.go を作らず、同じ states_test の frame_bench_test.go へ移す。
両者は vrt.InitReplayWorld + es.Init の設定も builders リストも同一だったので、ビルダーを
パッケージ変数 benchDungeons に集約して重複を消す。挙動は変えない。
