+++ internal/states/frame_bench_test.go
@@ -10,10 +10,22 @@ import (
 	es "github.com/kijimaD/ruins/internal/engine/states"
 	"github.com/kijimaD/ruins/internal/mapplanner"
 	gs "github.com/kijimaD/ruins/internal/states"
+	"github.com/kijimaD/ruins/internal/systems"
 	"github.com/kijimaD/ruins/internal/vrt"
+	"github.com/kijimaD/ruins/internal/world/query"
 	"github.com/mlange-42/ark/ecs"
 	"github.com/stretchr/testify/require"
 )
+
+// benchDungeons はフレーム・ターンのベンチで使う実ダンジョンのビルダー。
+var benchDungeons = []struct {
+	name string
+	bt   mapplanner.PlannerType
+}{
+	{"SmallRoom", mapplanner.PlannerTypeSmallRoom},
+	{"BigRoom", mapplanner.PlannerTypeBigRoom},
+	{"Cave", mapplanner.PlannerTypeCave},
+}
 
 // BenchmarkDungeonFrame は実ダンジョンの1フレームを Update / Draw に分けて実時間を計測する。
 //
@@ -25,16 +37,7 @@ import (
 // 通常プレイで描画が占める割合を示す。AIスパイク自体は BenchmarkProcessAll で別途計測している。
 // custom metric "gridEnts" は GridElement を持つエンティティ数（≒タイル数）で、Draw のスケール要因。
 func BenchmarkDungeonFrame(b *testing.B) {
-	builders := []struct {
-		name string
-		bt   mapplanner.PlannerType
-	}{
-		{"SmallRoom", mapplanner.PlannerTypeSmallRoom},
-		{"BigRoom", mapplanner.PlannerTypeBigRoom},
-		{"Cave", mapplanner.PlannerTypeCave},
-	}
-
-	for _, bld := range builders {
+	for _, bld := range benchDungeons {
 		world := vrt.InitReplayWorld(b)
 		sm, err := es.Init(&gs.DungeonState{Depth: 1, DefinitionName: dungeon.DungeonDebug.Name(), BuilderType: bld.bt}, world)
 		require.NoError(b, err)
@@ -58,6 +61,33 @@ func BenchmarkDungeonFrame(b *testing.B) {
 				require.NoError(b, sm.Draw(world, screen))
 			}
 			b.ReportMetric(gridEnts, "gridEnts")
+		})
+	}
+}
+
+// BenchmarkTurn は実ダンジョンで1ゲームターン全体の実時間を測る。vrt.InitReplayWorld と es.Init で
+// 本番のシステム登録込みのワールドを組み、TurnSystem を Phase=AI にして Update を2回回すと、
+// AI フェーズと End フェーズが走って1ターンが確定する。
+func BenchmarkTurn(b *testing.B) {
+	for _, bld := range benchDungeons {
+		world := vrt.InitReplayWorld(b)
+		sm, err := es.Init(&gs.DungeonState{Depth: 1, DefinitionName: dungeon.DungeonDebug.Name(), BuilderType: bld.bt}, world)
+		require.NoError(b, err)
+		// 数フレーム回して世界を落ち着かせてから測る
+		for range 5 {
+			require.NoError(b, sm.Update(world))
+		}
+
+		turnState := query.GetTurnState(world)
+		sys := &systems.TurnSystem{}
+
+		b.Run(bld.name, func(b *testing.B) {
+			// 反復ごとにワールドが1ターン進み状態は累積する。実プレイの流れを測るため毎回リセットしない
+			for range b.N {
+				turnState.Phase = gc.TurnPhaseAI
+				require.NoError(b, sys.Update(world)) // AI フェーズ -> End へ
+				require.NoError(b, sys.Update(world)) // End フェーズ -> Player へ。ターンが1つ確定する
+			}
 		})
 	}
 }
